backend/requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pandas
python-dotenv
playwright
lxml
html5lib
fastapi
uvicorn
pydantic
requests
